Fixed a SEGV error when using nanomsg on CentOS5/RHEL5
Older versions of glibc have a problem with getaddrinfo_a and the size of
stack it allocates for the thread that performs the lookup. This causes a
process built with nanomsg to crash when it needs to do a DNS lookup.

This fixes this issue by allowing the use of getaddrinfo_a to be disabled
at configuration time by specifying '--disable-getaddrinfo_a'.