fixes #533 nn_reallocmsg(p, -1000) causes an access violation on Windows
The test tried to use a fairly large negative number, which
evaluates to a positive big number on some platforms.  These
platforms may do various bad things with this result, including
throw assertions, etc.

Using a negative number smaller than the header size, we can
test for overflow of the unsigned size_t, which is what the test
is about.